final refactor (i hope). split up fmt / int_fmt determination to allow override of fmt without overriding int_fmt (useful for unifont.cpp). this changed the signature of ogl_tex_upload. better explained flags in OglTex and warn_if_uploaded.
also a few improvements / stomped on some warnings.

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@2809 3db68df2-c116-0410-a063-a993310a9797

+++ source/ps/ProfileViewer.cpp
@@ -62,7 +62,7 @@ void RenderProfile()
 
 	int estimate_height;
 	 
-	estimate_height = 6 + currentNode->GetChildren()->size() + currentNode->GetScriptChildren()->size();
+	estimate_height = 6 + (int)currentNode->GetChildren()->size() + (int)currentNode->GetScriptChildren()->size();
 	estimate_height = 20*estimate_height;
 	
 	glDisable(GL_TEXTURE_2D);
